+++ packages/flutter_watchos/example/watchos/Runner/GeneratedPluginRegistrant.m
@@ -22,6 +22,10 @@ extern void flutter_watchos_screen_width(void);
 extern void flutter_watchos_screen_height(void);
 extern void flutter_watchos_screen_scale(void);
 extern void flutter_watchos_play_haptic(void);
+extern void flutter_watchos_crown_mode(void);
+extern void flutter_watchos_crown_set_mode(void);
+extern void flutter_watchos_crown_push_delta(void);
+extern void flutter_watchos_crown_consume_delta(void);
 
 @implementation GeneratedPluginRegistrant
 
@@ -39,6 +43,10 @@ extern void flutter_watchos_play_haptic(void);
     (const void *)&flutter_watchos_screen_height,
     (const void *)&flutter_watchos_screen_scale,
     (const void *)&flutter_watchos_play_haptic,
+    (const void *)&flutter_watchos_crown_mode,
+    (const void *)&flutter_watchos_crown_set_mode,
+    (const void *)&flutter_watchos_crown_push_delta,
+    (const void *)&flutter_watchos_crown_consume_delta,
   };
   for (unsigned long _i = 0;
        _i < sizeof(_flutterWatchosFfiForcedReferences) / sizeof(_flutterWatchosFfiForcedReferences[0]);